Update Elliott.json
Fixed some run-on sentences and minor errors in grammar. Let me know if you have any questions for any of my edits (ping me @Arknir on discord). You do not need to accept all of the changes. Very good work!
Comments on some changes:
Line 23: Changed "couldn't" to "cannot". "Couldn't" is in the past tense, but Elliot seems to be speaking in the present tense. "Cannot" would seem better for this situation. The wording for "anything happening to you" seems okay, but you could add "bad" or "the thought of anything bad happening to you" 
Line 26: Changed "by Demetrius and Robin's home" to "mountain lake" for simplicity

Suggestions for edits by Arknir (has not been implemented, needs approval!):
Line 13: "It isn't the most beautiful one, but hey, still cool." ---> "Not my best catch, but it's still a keeper."
Line 19: "I wonder, how can a place that's so thriving also be so peaceful?" ---> "I wonder... how can a place that's thriving with so much life be so peaceful?"
Line 21: "You've started a fire in me that will not die," ---> "You've ignited a flame within me that will not burn out," 
Line 23: "I'm not sure what monsters we’ll face today, and I cannot bear anything happening to you!$2" ---> I'm not sure what monsters we’ll face today, and I cannot bear the thought of anything bad happening to you!$2",
Line 31: "Most of all, a real community, which is desperately lacking in so much of our modern world!" ---> "Most important of all, a loving community built on trust and friendships... Something that most places in our world desperately lack."